feat: hooks integration, wiki example, and i64 migration (#29)
* feat(hooks): repository macro integration, i64 migration, conn in after_* hooks

- Parse `hooks = Type` in #[repository] and generate transactional
  CRUD lifecycle (before_* → persist → after_* → commit → after_commit)
- Generate Patch<T> UpdateX, ModelField enum, DraftExt per-model
- Migrate primary keys from i32 to i64 across framework and examples
- Add conn parameter to after_* hooks for transactional side-effects
- Add compile-pass/fail tests, hooks_lifecycle, db_hooks_lifecycle


* feat(wiki): add wiki example with mutation hooks and revision audit

Showcases the hooks framework with a real use case:
- PageHooks: before_create auto-slugs, after_create/update inserts
  revision rows inside the same transaction (using conn param)
- Maud + HTMX UI with page list, show, edit, and revision history
- Docker Compose for local Postgres, embedded migrations


* docs: add implementation plans and architecture design document


* fix(ci): gate testcontainers tests behind #[ignore] for non-Docker runners

macOS and Windows CI runners lack Docker, causing db_hooks_lifecycle
tests to panic on SocketNotFoundError. Mark all 5 tests with
#[ignore = "requires Docker"] and add a Linux-only CI step that
runs ignored tests where Docker is available.


* fix(ci): scope --ignored to autumn-web db tests only

cargo test --workspace -- --ignored was running ALL ignored tests
including the autumn-cli e2e test which fails due to patch resolution.
Scope to just the db_hooks_lifecycle test file.


---------

+++ examples/blog/src/models.rs
@@ -10,7 +10,7 @@ use crate::schema::posts;
 #[diesel(table_name = posts)]
 #[diesel(check_for_backend(diesel::pg::Pg))]
 pub struct Post {
-    pub id: i32,
+    pub id: i64,
     pub title: String,
     pub slug: String,
     pub body: String,
@@ -40,7 +40,7 @@ impl Post {
     }
 
     /// Find a single post by ID, returning 404 if not found.
-    pub async fn find(id: i32, db: &mut AsyncPgConnection) -> AutumnResult<Self> {
+    pub async fn find(id: i64, db: &mut AsyncPgConnection) -> AutumnResult<Self> {
         posts::table
             .find(id)
             .select(Self::as_select())
